incorporate AppointmentSlidingPanel from main
