+++ frontend/src/router.ts
@@ -10,8 +10,8 @@ import { useUserStore } from '@/stores/user-store';
 import LogoutView from '@/views/LogoutView.vue';
 
 // lazy loaded components
-const ContactView = defineAsyncComponent(() => import('@/views/ContactView.vue'));
 const BookingsView = defineAsyncComponent(() => import('@/views/BookingsView/index.vue'));
+const ContactView = defineAsyncComponent(() => import('@/views/ContactView.vue'));
 const SettingsView = defineAsyncComponent(() => import('@/views/SettingsView.vue'));
 const ProfileView = defineAsyncComponent(() => import('@/views/ProfileView.vue'));
 const LegalView = defineAsyncComponent(() => import('@/views/LegalView.vue'));
@@ -112,7 +112,7 @@ const routes: RouteRecordRaw[] = [
     redirect: { name: 'bookings' },
   },
   {
-    path: '/bookings/:view?/:slug?',
+    path: '/bookings/:slug?',
     name: 'bookings',
     component: BookingsView,
     meta: {
