+++ frontend/src/views/BookingsView/components/AppointmentSlidingPanel.vue
